feat(cli): add daemon output handling and error management
- Introduce daemon_output module to parse and render daemon JSON messages
  streaming output to CLI stdout/stderr.
- Add AppError enum for structured CLI error handling with detailed variants.
- Migrate related helper functions to runtime_utils module.
- Modify main and lib to support terminal detection for output format auto-selection.
- Refactor tests and output rendering to improve error and output handling.
- Update documentation to clarify output auto-selection logic based on TTY.
