+++ crates/weaver-cli/src/lib.rs
@@ -5,20 +5,19 @@
 //! to be exercised both from the binary entrypoint and from tests where
 //! configuration loading and IO streams can be substituted.
 
-use std::ffi::{OsStr, OsString};
-use std::io::{self, IsTerminal, Write};
-use std::process::ExitCode;
-use std::sync::Arc;
-
 use clap::{Parser, Subcommand};
-use serde::Deserialize;
-use thiserror::Error;
-use weaver_config::{CapabilityMatrix, Config};
+use std::ffi::{OsStr, OsString};
+use std::io::Write;
+use std::process::ExitCode;
+use weaver_config::Config;
 
 mod command;
 mod config;
+mod daemon_output;
+mod errors;
 mod lifecycle;
 pub mod output;
+mod runtime_utils;
 mod transport;
 
 #[cfg(test)]
@@ -26,11 +25,15 @@ pub(crate) use command::CommandDescriptor;
 pub(crate) use command::{CommandInvocation, CommandRequest};
 use config::{ConfigArgumentSplit, split_config_arguments};
 pub(crate) use config::{ConfigLoader, OrthoConfigLoader};
+pub(crate) use daemon_output::{OutputSettings, read_daemon_messages};
+pub(crate) use errors::{AppError, is_daemon_not_running};
 use lifecycle::{
     LifecycleContext, LifecycleError, LifecycleInvocation, LifecycleOutput, SystemLifecycle,
     try_auto_start_daemon,
 };
 pub use output::{OutputContext, OutputFormat, ResolvedOutputFormat, render_human_output};
+use runtime_utils::emit_capabilities;
+pub(crate) use runtime_utils::exit_code_from_status;
 use transport::connect;
 /// CLI flags recognised by the configuration loader.
 ///
@@ -44,7 +47,7 @@ const CONFIG_CLI_FLAGS: &[&str] = &[
     "--log-format",
     "--capability-overrides",
 ];
-const EMPTY_LINE_LIMIT: usize = 10;
+pub(crate) const EMPTY_LINE_LIMIT: usize = 10;
 
 /// Bundles the IO streams provided to the CLI runtime.
 ///
@@ -59,11 +62,11 @@ pub(crate) struct IoStreams<'a, W: Write, E: Write> {
 }
 
 impl<'a, W: Write, E: Write> IoStreams<'a, W, E> {
-    pub(crate) fn new(stdout: &'a mut W, stderr: &'a mut E) -> Self {
+    pub(crate) fn new(stdout: &'a mut W, stderr: &'a mut E, stdout_is_terminal: bool) -> Self {
         Self {
             stdout,
             stderr,
-            stdout_is_terminal: io::stdout().is_terminal(),
+            stdout_is_terminal,
         }
     }
 
@@ -188,13 +191,13 @@ where
 
 /// Runs the CLI using the provided arguments and IO handles.
 #[must_use]
-pub fn run<I, W, E>(args: I, stdout: &mut W, stderr: &mut E) -> ExitCode
+pub fn run<I, W, E>(args: I, stdout: &mut W, stderr: &mut E, stdout_is_terminal: bool) -> ExitCode
 where
     I: IntoIterator<Item = OsString>,
     W: Write,
     E: Write,
 {
-    let mut io = IoStreams::new(stdout, stderr);
+    let mut io = IoStreams::new(stdout, stderr, stdout_is_terminal);
     run_with_loader(args, &mut io, &OrthoConfigLoader)
 }
 
@@ -334,109 +337,6 @@ where
         .run_with_handler(args, handler)
 }
 
-fn emit_capabilities<W>(config: &Config, stdout: &mut W) -> Result<(), AppError>
-where
-    W: Write,
-{
-    let matrix: CapabilityMatrix = config.capability_matrix();
-    serde_json::to_writer_pretty(&mut *stdout, &matrix).map_err(AppError::SerialiseCapabilities)?;
-    stdout
-        .write_all(b"\n")
-        .map_err(AppError::EmitCapabilities)?;
-    stdout.flush().map_err(AppError::EmitCapabilities)
-}
-
-fn exit_code_from_status(status: i32) -> ExitCode {
-    if (0..=255).contains(&status) {
-        ExitCode::from(status as u8)
-    } else {
-        ExitCode::FAILURE
-    }
-}
-
-/// Determines whether an error indicates the daemon is not running.
-///
-/// Returns true for connection-refused, socket-not-found, and address-unavailable
-/// errors, which typically indicate the daemon process is not listening.
-fn is_daemon_not_running(error: &AppError) -> bool {
-    match error {
-        AppError::Connect { source, .. } => matches!(
-            source.kind(),
-            io::ErrorKind::ConnectionRefused
-                | io::ErrorKind::NotFound
-                | io::ErrorKind::AddrNotAvailable
-        ),
-        _ => false,
-    }
-}
-
-/// Settings for rendering daemon output.
-pub(crate) struct OutputSettings<'a> {
-    pub(crate) format: ResolvedOutputFormat,
-    pub(crate) context: &'a OutputContext,
-}
-
-fn read_daemon_messages<R, W, E>(
-    connection: &mut R,
-    io: &mut IoStreams<'_, W, E>,
-    settings: OutputSettings<'_>,
-) -> Result<i32, AppError>
-where
-    R: io::Read,
-    W: Write,
-    E: Write,
-{
-    use std::io::BufRead;
-
-    let mut reader = io::BufReader::new(connection);
-    let mut line = String::new();
-    let mut exit_status: Option<i32> = None;
-    let mut consecutive_empty_lines = 0;
-
-    while reader
-        .read_line(&mut line)
-        .map_err(AppError::ReadResponse)?
-        != 0
-    {
-        if line.trim().is_empty() {
-            consecutive_empty_lines += 1;
-            if consecutive_empty_lines >= EMPTY_LINE_LIMIT {
-                writeln!(
-                    io.stderr,
-                    "Warning: received {EMPTY_LINE_LIMIT} consecutive empty lines from daemon; aborting."
-                )
-                    .map_err(AppError::ForwardResponse)?;
-                break;
-            }
-            line.clear();
-            continue;
-        }
-        consecutive_empty_lines = 0;
-        let message: DaemonMessage = serde_json::from_str(&line).map_err(AppError::ParseMessage)?;
-        match message {
-            DaemonMessage::Stream { stream, data } => {
-                let rendered = match settings.format {
-                    ResolvedOutputFormat::Human => render_human_output(settings.context, &data),
-                    ResolvedOutputFormat::Json => None,
-                };
-                let payload = rendered.as_deref().unwrap_or(&data);
-                match stream {
-                    StreamTarget::Stdout => io.stdout.write_all(payload.as_bytes()),
-                    StreamTarget::Stderr => io.stderr.write_all(payload.as_bytes()),
-                }
-                .map_err(AppError::ForwardResponse)?;
-            }
-            DaemonMessage::Exit { status } => exit_status = Some(status),
-        }
-        line.clear();
-    }
-
-    io.stdout.flush().map_err(AppError::ForwardResponse)?;
-    io.stderr.flush().map_err(AppError::ForwardResponse)?;
-
-    exit_status.ok_or(AppError::MissingExit)
-}
-
 #[derive(Parser, Debug)]
 #[command(
     name = "weaver",
@@ -486,57 +386,6 @@ enum DaemonAction {
     Stop,
     /// Prints daemon health information.
     Status,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-enum DaemonMessage {
-    Stream { stream: StreamTarget, data: String },
-    Exit { status: i32 },
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum StreamTarget {
-    Stdout,
-    Stderr,
-}
-
-#[derive(Debug, Error)]
-enum AppError {
-    #[error("failed to load configuration: {0}")]
-    LoadConfiguration(Arc<ortho_config::OrthoError>),
-    #[error("{0}")]
-    CliUsage(clap::Error),
-    #[error("the command domain must be provided")]
-    MissingDomain,
-    #[error("the command operation must be provided")]
-    MissingOperation,
-    #[error("failed to resolve daemon address {endpoint}: {source}")]
-    Resolve { endpoint: String, source: io::Error },
-    #[error("failed to connect to daemon at {endpoint}: {source}")]
-    Connect { endpoint: String, source: io::Error },
-    #[cfg(not(unix))]
-    #[error("platform does not support Unix sockets: {0}")]
-    UnsupportedUnixTransport(String),
-    #[error("failed to serialise command request: {0}")]
-    SerialiseRequest(serde_json::Error),
-    #[error("failed to send request to daemon: {0}")]
-    SendRequest(io::Error),
-    #[error("failed to read response from daemon: {0}")]
-    ReadResponse(io::Error),
-    #[error("failed to parse daemon message: {0}")]
-    ParseMessage(serde_json::Error),
-    #[error("failed to forward daemon output: {0}")]
-    ForwardResponse(io::Error),
-    #[error("daemon closed the stream without sending an exit status")]
-    MissingExit,
-    #[error("failed to serialise capability matrix: {0}")]
-    SerialiseCapabilities(serde_json::Error),
-    #[error("failed to emit capabilities: {0}")]
-    EmitCapabilities(io::Error),
-    #[error("daemon lifecycle command failed: {0}")]
-    Lifecycle(#[from] LifecycleError),
 }
 
 #[cfg(test)]
