+++ crates/weaver-cli/src/main.rs
@@ -4,11 +4,17 @@
 //! processes command-line arguments, negotiates capability output, and streams
 //! JSONL requests to the configured daemon transport.
 
-use std::io::{self, StderrLock, StdoutLock};
+use std::io::{self, IsTerminal, StderrLock, StdoutLock};
 use std::process::ExitCode;
 
 fn main() -> ExitCode {
+    let stdout_is_terminal = io::stdout().is_terminal();
     let mut stdout: StdoutLock<'_> = io::stdout().lock();
     let mut stderr: StderrLock<'_> = io::stderr().lock();
-    weaver_cli::run(std::env::args_os(), &mut stdout, &mut stderr)
+    weaver_cli::run(
+        std::env::args_os(),
+        &mut stdout,
+        &mut stderr,
+        stdout_is_terminal,
+    )
 }
